half-moved ULF wave chapter.
the two halves of the ULF wave chapter -- background on the waves, and computation of the dispersion relation -- don't really fit together. The former makes more sense in teh intro chapter, and allows us to finish the first chapter talking about goals of this work. the latter... might work with the numerical model? it's unclear. at the moment, it's an awkward, short chapter.