rather than all subsection, why not go NO subsection?
we can still have the entire thesis at the same level... it's just one less level. we add a little fluff intro with the structure of the work, then break each section into its own chapter. this gets rid of the useless blurb that we've been struggling to write at the top of each of the five super-chapters. it also makes the observations fit better... they aren't big enough to be their own chapter (on the level of dispersion relation + model + inertia) but they are on par with one of those things.